+++ docs/changes.rst
@@ -12,6 +12,8 @@ To be released.
   :exc:`AttributeError` when the given HTML contains ``<link>`` to
   both :mimetype:`application/atom+xml` and :mimetype:`application/rss+xml`.
   [:issue:`40`]
+- Fill ``<title>`` to ``<description>`` if there's no ``<title>``
+  (:mod:`~libearth.parser.rss2`).
 
 
 Version 0.1.1